Updated network-based skip rules for FeesReimbursementPool configuration
The FeesReimbursementPool address in the BitcoinDepositor can be set by
the script only on Hardhat network. On all other networks it has to be
done manually by the governance.

Accepting ownership of FeesReimbursementPool can only happen on Hardhat
network. On all other networks, the governance address need to call this
function and the governance is either external EOA (Sepolia) or a
multisig (Ethereum mainnet).