Initial implementation of the MidasAllocator contract for deposits allocation
The MidasAllocator contract is a simple allocator that deposits tBTC to the
Midas Vault and holds the shares minted by the Vault.

The Vault is expected to be implemented by the Midas team according to the
IVault interface, to wrap their own vault contracts. As their setup uses two
distinct vaults for deposits and redemptions, in addition to a separate shares
token contract.

This implementation doesn't cover the withdrawal functionality, as it will
be built on top of the queued withdrawals functionality.