Add migration feature to stBTC contract
The stBTC contract is enhanced with a migration feature that allows to migrate
the deposit of the owner to a new ERC-4626 contract.

Once the migration is started, the funds are released from the MezoAllocator
and routed to the stBTC contract.

The MezoAllocator contract is deauthorized from accessing the assets of the
stBTC contract.

During the migration, new deposits are not allowed.

The owner of the stBTC contract can migrate the deposits of the depositors
without their approval. The migration is done by burning the shares that are
being migrated and depositing the assets to the new contract.

The migration takes into account the shares that were minted as debt, excluding
them from the assets that are being migrated.