Fix assets calculation in migrateDeposit function
The function stBTC.migrateDeposit() calculates the amount of assets to
be migrated by converting to assets the total share balance of the owner.

However, if the owner has less withdrawable shares than its balance, the
amount of burnt shares is lowered. As a result a migration could migrate
an amount of assets inflated in comparison to the amount of burnt shares.