Expect exit fee to be passed to the withdrawal queue
The exit fee will be calculated in acreBTC contract and passed in the request
to the withdrawal queue. When finalizing the request, the exit fee will be
taken matching the fee that was set on the initial request.

This change extracts exit fee logic to the withdrawal queue and makes it
consistent across regular redemption request and redemption to bitcoin.