Implement BitcoinRedeemerV2 contract
The contract is used to redeem acreBTC shares with async redemption from
Acre Vault and bridging to Bitcoin.

This contract is related to the BitcoinRedeemer contract known from stBTC,
but adjusted to support new redemption interfaces for acreBTC.

Requests for redemptions are expected to be submitted to the Acre Vault
with `approveAndCall` functionality, pointing to this contract. This path
will be used by Ledger Live applications for clear signing the requests.