Validate minimum redemption amount for bridging to Bitcoin
Refactor WithdrawalQueue to validate minimum redemption amount for bridging to Bitcoin.

The value is checked by the tBTC Bridge contract on the finalization of the redemption request.
But we need to validate it when the request is initiated, so the funds are not locked in the WithdrawalQueue
due to using too small amount for redemption request.