Allow zero address for withdrawal queue address
Allow zero address for withdrawal queue address in MidasAllocator contract.
This makes it consistent with the acreBTC contract, where we also allow zero address.

In this case the withdrawal queue is used only in `withdrawShares` function,
to check the msg.sender against it. It's fine to allow zero as the zero address
can't call the function.