Removed reimbursement mechanism from BitcoinDepositorV2
The `tbtcAmount` returned from `_finalizeDeposit(depositKey)` is always lower
than the `initialAmount` so we would reimburse even if the amount of minted tBTC
was equal to the `initialAmount`. As a result, we would keep moving funds from
`FeesReimbursementPool` to `BitcoinDepositorV2` with each individual deposit.

The approximation done by `_finalizeDeposit(depositKey)` is the worst-case
scenario, and the worst-case scenario is far from reality, given that the
`depositTxMaxFee` must be sufficiently high.

I analyzed all BTC deposit transactions to Mezo and all of them were handled by
tBTC optimistic minting. Even if they were not done by optimistic minting, the
average sweep cost per deposit was $0.37. The `depositTxMaxFee` is currently set
to 0.001 BTC, so $113 with today's price. The difference is significant.

It is impossible to say whether tBTC was minted via optimistic minting or
sweeping. It's also not possible to distinguish individual tBTC minting
operations to the `BitcoinDepositorV2` contract with concurrent deposits.

What we can do however is to trust all Acre deposits will be optimistically
minted and if not, leave some tBTC safety buffer in the depositor contract.

This change removes fee reimbursement pool completely, and the code now trusts
tBTC bridge to mint tokens 1:1 to the originally deposited amount in BTC.

Worth noting, I also fixed how the minted amount is calculated in the stub tBTC
Vault contract. The previous implementation wasn't correct as it was also
subtracting maximum deposit swap fee - this is NOT how tBTC vault contract
works.