Track redemption key to withdraw id
There can be multiple withdraw entities with the same redemption key,
but only one redemption can be in the "pending" state. As a result,
there can be only one redemption with the same redemption key and the
same withdraw ID. It means that on the tBTC side, first, the redemption
is initialized and a redemption key is generated. Until this redemption
is processed (i.e., finalized), it is not possible to create another
redemption with the same key. To determine whether a given redemption
has been processed, we track the `submitRedemptionProofCall` function
from the `TbtcBridge` contract. However, the data emitted by this
function does not allow us to link the redemption to the withdraw ID
generated by the contract. Therefore, we need to figure out a way to
associate the withdraw ID with the redemption key — and the
`RedemptionKeyToPendingWithdrawal` entity serves that purpose.