Fix the withdrawal initialization
In the previous implementation, we passed the parameter
`redeemerOutputScript` incorrectly. It was not prepended with its
length, which caused the tBTC redemption process for this withdrawal
could not be executed. The `redeemerOutputScript` must be a
length-prefixed output script. To ensure the correct format, we now
build the data using the tBTC v2 SDK, which provides the proper format.