Update `WithdrawalStatusBanner` component
In the previous implementation if a withdrawal hasn't been processed
within 72 hours, the user saw negative values under the progress bar.
Such situations may occur if something gets delayed on the system’s
side. For this case let's just display `Est. duration 72 hours`. This is
not optimal and should be reconsidered and improved to find a clearer
and more user-friendly way to display it.